requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
mysql-connector-python==9.3.0
psycopg2-binary==2.9.9
fpdf2==2.8.3
Pillow==11.2.1
python-dotenv==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
SQLAlchemy==2.0.41
greenlet==3.2.3
defusedxml==0.7.1
colorama==0.4.6
gunicorn==21.2.0 